added linear sieve

+++ linear-sieve.cpp
@@ -5,8 +5,8 @@ vector<int> pr;
 for (int i=2; i<=N; ++i) {
 	if (lp[i] == 0) {
 		lp[i] = i;
-		pr.push_back (i);
+		pr.push_back(i);
 	}
-	for (int j=0; j<(int)pr.size() && pr[j]<=lp[i] && i*pr[j]<=N; ++j)
+	for (int j=0; j<pr.size() && pr[j]<=lp[i] && i*pr[j]<=N; ++j)
 		lp[i * pr[j]] = pr[j];
 }
